requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
requests==2.32.4
apscheduler==3.11.0
python-dotenv==1.1.1
